requirements.txt
Django==2.2.10
djangorestframework==3.12.2
psycopg2-binary>=2.8
django-rest-registration==0.5.6
django-filter==2.4.0